requirements.txt
requests>=2.32.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
PyYAML>=6.0.1
pdfplumber>=0.11.4
pypdf>=4.3.1
python-dateutil>=2.9.0
